+++ mall-portal/src/main/java/org/example/modules/product/entity/SkuStockEntity.java
@@ -9,6 +9,7 @@ import lombok.NoArgsConstructor;
 import org.example.common.core.base.CommonEntity;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -43,7 +44,7 @@ public class SkuStockEntity extends CommonEntity<SkuStockEntity> implements Seri
      * 价格
      */
     @Schema(name = "price", description = "价格")
-    private Double price;
+    private BigDecimal price;
     /**
      * 库存
      */
@@ -68,7 +69,7 @@ public class SkuStockEntity extends CommonEntity<SkuStockEntity> implements Seri
      * 单品促销价格
      */
     @Schema(name = "promotionPrice", description = "单品促销价格")
-    private Double promotionPrice;
+    private BigDecimal promotionPrice;
     /**
      * 锁定库存
      */
