+++ mall-system/src/main/java/org/example/modules/product/entity/dto/SkuStockDto.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Dou-Fu-10 2023-07-15 11:35:50
  *
@@ -29,7 +31,7 @@ public class SkuStockDto {
     /**
      * 价格
      */
-    private Double price;
+    private BigDecimal price;
     /**
      * 库存
      */
@@ -49,7 +51,7 @@ public class SkuStockDto {
     /**
      * 单品促销价格
      */
-    private Double promotionPrice;
+    private BigDecimal promotionPrice;
     /**
      * 锁定库存
      */
