+++ mall-system/src/main/java/org/example/modules/product/entity/vo/ProductVo.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import java.util.Set;
@@ -80,7 +81,7 @@ public class ProductVo {
     /**
      * 价格
      */
-    private Double price;
+    private BigDecimal price;
     /**
      * 副标题
      */
@@ -92,7 +93,7 @@ public class ProductVo {
     /**
      * 市场价
      */
-    private Double originalPrice;
+    private BigDecimal originalPrice;
     /**
      * 库存
      */
@@ -108,7 +109,7 @@ public class ProductVo {
     /**
      * 商品重量，默认为克
      */
-    private Double weight;
+    private BigDecimal weight;
     /**
      * 以逗号分割的产品服务：1->无忧退货；2->快速退款；3->免费包邮
      */
